fixed resize() logic

+++ proj1/deque/ArrayDequeTest.java
@@ -148,12 +148,12 @@ public class ArrayDequeTest {
             lld1.addLast(i);
         }
 
-        for (double i = 0; i < 500000; i++) {
-            assertEquals("Should have the same value", i, (double) lld1.removeFirst(), 0.0);
+        for (int i = 0; i < 500000; i++) {
+            assertEquals("Should have the same value", i,  lld1.removeFirst(), 0.0);
         }
 
-        for (double i = 999999; i > 500000; i--) {
-            assertEquals("Should have the same value", i, (double) lld1.removeLast(), 0.0);
+        for (int i = 999999; i > 500000; i--) {
+            assertEquals("Should have the same value", i,  lld1.removeLast(), 0.0);
         }
 
     }
